requirements_v2.txt
# Core Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# YOLOv8 and Deep Learning
ultralytics==8.2.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
scikit-image==0.22.0
albumentations==1.3.1

# Image Processing
Pillow==10.0.1
pdf2image==1.16.3
imageio==2.31.5

# Scientific Computing
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2

# Data Augmentation and Processing
imgaug==0.4.0
tqdm==4.66.1

# Database
sqlite3-python==1.0.0

# Server
gunicorn==21.2.0
python-multipart==0.0.6

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0

# Monitoring and Logging
tensorboard==2.14.0
wandb==0.15.12

# Development (optional)
ipython==8.17.2
jupyter==1.0.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
